Added a feature or two
I thought that replacing `(N, W, E, S)` with `(N+W+E+S)` would solve the
display program... It didn't.

Replaced the canvas with a label. It's easier to work with and plays
nicely with configuration.

I did have for a while a "default.png" file that came with my script. I
immediately realized that, that wasn't necessary. It is now created upon
start of the script and WILL be deleted upon close.

Changed 'EVENT NAME' to 'DEFAULT EVENT NAME', just for clarification.

I used ".configure" of Tkinter to change the label image.