Lots of changes and improvements
The name of the main file has been changed. Sorry, I know this makes
changes difficult to identify but it's in anticipation for quantifying
the project at a later date.

Added a context menu instead Password Change. This will allow the
program to act in "context" of an event based on your current focus.

Added an options tab (just for debugging at the moment) but more
features will be added later.